Transitory refactoring  of daemon package and modules and its internal reference  to rearrange the daemon files in a sub-package ('gnr.web.daemon'), which will allow easier plugging via entrypoints of external packages.
