+++ gnrpy/gnr/web/daemon/handler.py
@@ -20,7 +20,7 @@ from gnr.core.gnrbag import Bag
 from gnr.core.gnrsys import expandpath
 from gnr.core.gnrconfig import gnrConfigPath
 from gnr.app.gnrdeploy import PathResolver
-from gnr.web.gnrdaemonprocesses import GnrCronHandler, GnrDaemonServiceManager
+from gnr.web.daemon.processes import GnrCronHandler, GnrDaemonServiceManager
 from gnr.web import gnrtask
 from gnr.web import logger
