+++ gnrpy/tests/web/gnrdaemon.py
@@ -1,0 +1,1 @@
+import gnr.web.daemon  # noqa: F401
