--- gnrpy/tests/web/gnrdaemonhandler_test.py
@@ -1,1 +1,0 @@
-import gnr.web.gnrdaemonhandler  # noqa: F401
