--- gnrpy/tests/web/gnrdaemonprocesses_test.py
@@ -1,1 +1,0 @@
-import gnr.web.gnrdaemonprocesses  # noqa: F401
